refactor: split different class for different workflow
