+++ zeebe-job-worker/src/main/kotlin/workflowmanager/Deployments.kt
@@ -1,8 +1,6 @@
-package processmanager
+package workflowmanager
 
 import io.camunda.zeebe.spring.client.annotation.Deployment
-import org.springframework.boot.SpringApplication
-import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.context.annotation.Configuration
 
 @Configuration
