+++ zeebe-job-worker/src/main/kotlin/workflowmanager/Main.kt
@@ -1,4 +1,4 @@
-package processmanager
+package workflowmanager
 
 import lombok.extern.slf4j.Slf4j
 import org.springframework.boot.autoconfigure.SpringBootApplication
@@ -6,7 +6,7 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
-@ConfigurationPropertiesScan("processmanager")
+@ConfigurationPropertiesScan("workflowmanager")
 @Slf4j
 open class Main
